--- src/main/kotlin/Cuenta.kt
@@ -1,32 +1,0 @@
-
-/*  Cuenta.kt tendra los atributos para poder manejar las cuentas de los usuarios y las
-    operaciones que se pueden realizar en ellas.
-
-    Posible atributos:
-    - Numero de cuenta
-    - Saldo
-    - Tipo de cuenta
-    - Fecha de creacion
-    - Fecha de modificacion
-    - Fecha de eliminacion
-    - Estado de la cuenta
-    - Usuario
-    - Transacciones
-
-    Posibles Operaciones:
-    - Deposito
-    - Retiro
-    - Transferencia
-    - Consulta de saldo
-    - Consulta de transacciones
-    - Consulta de estado de cuenta
-    - Consulta de fecha de creacion
-    - Consulta de fecha de modificacion
-    - Consulta de fecha de eliminacion
-    - Consulta de tipo de cuenta
-    - Consulta de usuario
-
-
-*/
-class Cuenta {
-}
