+++ src/main/kotlin/mx/com/team9/activities/Retiro.kt
@@ -1,3 +1,8 @@
+package mx.com.team9.activities
+
+import mx.com.team9.dataclase.Categoria
+import mx.com.team9.models.Cuenta
+import mx.com.team9.utils.TipoMovimiento
 import java.time.LocalDateTime
 /**
  * Clase que representa un retiro de una cuenta
@@ -13,4 +18,18 @@ class Retiro : Transaccion() {
         get() = TODO("Not yet implemented")
     override val categoria: Categoria
         get() = TODO("Not yet implemented")
+    override val tipo: TipoMovimiento
+        get() = TODO("Not yet implemented")
+
+    override fun movimiento(
+        monto: Double,
+        cuenta: Cuenta,
+        tipo: TipoMovimiento,
+        descripcion: String,
+        categoria: Categoria
+    ) {
+        TODO("Not yet implemented")
+    }
+
+
 }
