+++ src/main/kotlin/mx/com/team9/dataclase/Categoria.kt
@@ -1,3 +1,7 @@
+package mx.com.team9.dataclase
+
+import mx.com.team9.utils.TipoCategoria
+
 /**
  * Data class que representa una categoría de una transaccion
  */
