+++ src/main/kotlin/mx/com/team9/models/Reporte.kt
@@ -1,3 +1,5 @@
+package mx.com.team9.models
+
 // Clase que generarada por el usuario que genera los reportes de una cuenta
 class Reporte {
 }
