+++ src/main/kotlin/mx/com/team9/utils/TipoCategoria.kt
@@ -1,3 +1,5 @@
+package mx.com.team9.utils
+
 /**
  * Enumeración que representa los tipos de categorías que pueden tener las transacciones
  * @property HOGAR
