+++ src/main/kotlin/mx/com/team9/utils/TipoMovimiento.kt
@@ -1,0 +1,6 @@
+package mx.com.team9.utils
+
+enum class TipoMovimiento {
+    DEPOSITO,
+    RETIRO
+}
